backend/requirements.txt
pandas
numpy
faiss-cpu 
sentence-transformers 
tqdm
langgraph 
langchain 
python-dotenv
transformers 
torch
langchain-google-genai
torchvision
langchain-groq
groq
fastapi
uvicorn
pydantic
